Updates based on 2022-02-24_Instructions text file
For section 1.1. Added columns for
Number of unique firms, average number of year per firm
Round down to two decimal places when writing it to the table for the p10 of sales.

Clean size category construction from the balanced panel construction

Sections 1.6 and 1.7 are replaced with 1.12

Sections 1.8 and 1.9 y-axis labelling

Add graph - Share of firms by age before 1.14

Section 2.2 Germany and Netherlands would look better with zeroes and not scientific notation